+++ tests/test_telephony_devices_line_key_templates.py
@@ -3,12 +3,12 @@ Test Line key template operations
 """
 import asyncio
 import json
-from random import choice
+from time import sleep
 
 from tests.base import TestCaseWithLog, async_test, TestWithLocations
-from wxc_sdk.common import ApplyLineKeyTemplateAction
+from wxc_sdk.common import ApplyLineKeyTemplateAction, OwnerType
 from wxc_sdk.devices import Device, ProductType
-from wxc_sdk.telephony import NumberType, OwnerType, NumberListPhoneNumber, SupportedDevice
+from wxc_sdk.telephony import NumberType, NumberListPhoneNumber, SupportedDevice
 from wxc_sdk.telephony.devices import LineKeyTemplate, ProgrammableLineKey, LineKeyType
 
 
@@ -19,7 +19,7 @@ class TestLineKeyTemplate(TestCaseWithLog):
         """
         api = self.api.telephony.devices
         lkts = list(api.list_line_key_templates())
-        foo = 1
+        print(f'found {len(lkts)} line key templates')
 
     @async_test
     async def test_details(self):
@@ -31,7 +31,10 @@ class TestLineKeyTemplate(TestCaseWithLog):
         details = await asyncio.gather(*[api.line_key_template_details(template_id=lkt.id)
                                          for lkt in lkts],
                                        return_exceptions=True)
-        foo = 1
+        err = next((d for d in details if isinstance(d, Exception)), None)
+        if err:
+            raise err
+        print(f'for details for {len(lkts)} line key templates')
 
     @async_test
     async def test_create(self):
@@ -41,7 +44,7 @@ class TestLineKeyTemplate(TestCaseWithLog):
         # get some information 1st
         extensions, supported, devices, lkts = await asyncio.gather(
             self.async_api.telephony.phone_numbers(number_type=NumberType.extension,
-                                                           owner_type=OwnerType.people),
+                                                   owner_type=OwnerType.people),
             self.async_api.telephony.supported_devices(),
             self.async_api.devices.list(product_type=ProductType.phone),
             self.async_api.telephony.devices.list_line_key_templates())
@@ -54,7 +57,7 @@ class TestLineKeyTemplate(TestCaseWithLog):
         lkt_name = next(name
                         for i in range(1, 100)
                         if (name := f'test 8865 {i:03}') not in set(lkt.template_name
-                                                                for lkt in lkts))
+                                                                    for lkt in lkts))
         line_keys = ProgrammableLineKey.standard_plk_list(10)
         line_keys[5] = ProgrammableLineKey(line_key_index=6, line_key_type=LineKeyType.speed_dial,
                                            line_key_label='hgreen',
@@ -69,15 +72,77 @@ class TestLineKeyTemplate(TestCaseWithLog):
 
 
 class TestPreviewApply(TestWithLocations):
-    def test_preview_apply(self):
+    @async_test
+    async def test_preview_apply_default_template(self):
         """
         try to apply a line key template
         """
-        lkts = list(self.api.telephony.devices.list_line_key_templates())
-        if not lkts:
-            self.skipTest('No line key templates')
-        target_lkt = choice(lkts)
-        all_location_ids = [loc.location_id for loc in self.locations]
+        devices = await self.async_api.devices.list(product_type=ProductType.phone)
+        if not devices:
+            self.skipTest('No MPP devices')
+
+        # preview apply default template
         r = self.api.telephony.devices.preview_apply_line_key_template(
             action=ApplyLineKeyTemplateAction.apply_default_templates)
-        foo = 1
+        self.assertEqual(len(devices), r)
+
+
+class TestJobs(TestCaseWithLog):
+
+    @async_test
+    async def test_list_jobs(self):
+        """
+        list apply line key template jobs
+        """
+        jobs = await self.async_api.telephony.jobs.apply_line_key_templates.list_jobs()
+        print(f'got {len(jobs)} jobs')
+
+    @async_test
+    async def test_job_status(self):
+        """
+        get status of all jobs
+        """
+        jobs = await self.async_api.telephony.jobs.apply_line_key_templates.list_jobs()
+        details = await asyncio.gather(
+            *[self.async_api.telephony.jobs.apply_line_key_templates.job_status(job_id=job.id) for job in jobs],
+            return_exceptions=True)
+        err = next((d for d in details if isinstance(d, Exception)), None)
+        if err:
+            raise err
+        print(f'got status for {len(details)} jobs')
+
+    @async_test
+    async def test_job_errors(self):
+        """
+        get errors of all jobs
+        """
+        jobs = await self.async_api.telephony.jobs.apply_line_key_templates.list_jobs()
+        errors = await asyncio.gather(
+            *[self.async_api.telephony.jobs.apply_line_key_templates.job_errors(job_id=job.id) for job in jobs],
+            return_exceptions=True)
+        err = next((d for d in errors if isinstance(d, Exception)), None)
+        if err:
+            raise err
+        print(f'got errors for {len(errors)} jobs')
+
+    def test_apply_job(self):
+        """
+        Apply line key template job, default like key template
+        """
+        api = self.api.telephony.jobs.apply_line_key_templates
+        r = self.api.telephony.devices.preview_apply_line_key_template(
+            action=ApplyLineKeyTemplateAction.apply_default_templates,
+            exclude_devices_with_custom_layout=True)
+        job = api.apply(
+            action=ApplyLineKeyTemplateAction.apply_default_templates,
+            exclude_devices_with_custom_layout=True)
+        print(json.dumps(job.model_dump(mode='json', exclude_none=True, by_alias=True), indent=2))
+        # wait for job to complete
+        while True:
+            status = api.job_status(job_id=job.id)
+            print(json.dumps(status.model_dump(mode='json', exclude_none=True, by_alias=True), indent=2))
+            print()
+            if status.latest_execution_status == 'COMPLETED':
+                break
+            sleep(10)
+        self.assertEqual(r, status.updated_count)