+++ docs/changes.rst
@@ -22,6 +22,7 @@ Release history
 - fix: :attr:`wxc_sdk.telephony.callqueue.CallQueue.department`, optional
 - fix: :class:`wxc_sdk.common.OwnerType` needs to support PAGING_GROUP and GROUP_PAGING (inconsistent)
 - fix: undocumented attribute :attr:`wxc_sdk.telephone.voicemail_groups.VoicemailGroupDetail.time_zone`
+- feat: new API :attr:`wxc_sdk.telephony.jobs.JobsApi.apply_line_key_templates`
 
 1.16.1
 ------